✨ update | 카트 엔티티, 오더 엔티티 수정

+++ final_project_backend/src/merchandise/merchandise.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   HttpStatus,
   Injectable,
   NotFoundException,
@@ -16,6 +17,7 @@ import { FindAllmerchandiseDto } from './dto/find-merchandise.dto';
 import { UpdateMerchandiseDto } from './dto/update-merchandise.dto';
 import { title } from 'process';
 import * as Flatted from 'flatted';
+import { Manager } from 'src/admin/entities/manager.entity';
 
 @Injectable()
 export class MerchandiseService {
@@ -28,33 +30,40 @@ export class MerchandiseService {
     private readonly productRepository: Repository<Product>,
     @InjectRepository(MerchandiseOption)
     private readonly merchandiseOptionRepository: Repository<MerchandiseOption>,
+    @InjectRepository(Manager)
+    private readonly managerRepository: Repository<Manager>,
   ) {}
 
   // 상품 생성 API
-  async create(createMerchandiseDto: CreateMerchandiseDto) {
-    //유저 받아오기 추가 필요
+  async create(createMerchandiseDto: CreateMerchandiseDto, userId: number) {
     const { productId, imageUrl, option, optionPrice, ...merchandiseData } =
       createMerchandiseDto;
 
+    // 상점 유효성 체크
     const product = await this.productRepository.findOne({
       where: { id: productId },
     });
-
-    // 상점 id 체크
     if (!product) {
       throw new NotFoundException('상점이 존재하지 않습니다');
     }
 
+    // 옵션 + 옵션명 수 동일 체크
     if (option.length !== optionPrice.length) {
       throw new BadRequestException(
         '옵션명과 옵션 가격의 수가 일치하지 않습니다',
       );
     }
 
+    //매니저 정보 가져오기
+    const manager = await this.managerRepository.findOne({
+      where: { userId },
+    });
+
     // 상점 아이디가 있을 경우 생성
     const merchandise = await this.merchandiseRepository.create({
       ...merchandiseData,
       product,
+      manager,
     });
     const saveMerchandise = await this.merchandiseRepository.save(merchandise);
 
@@ -79,8 +88,8 @@ export class MerchandiseService {
       status: HttpStatus.CREATED,
       message: '상품 생성이 완료되었습니다',
       data: {
-        id: merchandise.id,
-        userId: '추가예정',
+        merchandiseId: merchandise.id,
+        managerId: manager.managerId,
         title: merchandise.title,
         thumbnail: merchandise.thumbnail,
         salesName: merchandise.salesName,
@@ -166,7 +175,11 @@ export class MerchandiseService {
   }
 
   // 상품 수정 API
-  async update(id: number, updateMerchandiseDto: UpdateMerchandiseDto) {
+  async update(
+    id: number,
+    updateMerchandiseDto: UpdateMerchandiseDto,
+    userId: number,
+  ) {
     const {
       title,
       thumbnail,
@@ -178,16 +191,19 @@ export class MerchandiseService {
       optionPrice,
     } = updateMerchandiseDto;
 
+    //상품 유효성 체크
     const merchandise = await this.merchandiseRepository.findOne({
       where: { id },
-      relations: ['merchandiseImage', 'merchandiseOption'],
+      relations: ['merchandiseImage', 'merchandiseOption', 'manager'],
     });
-
     if (!merchandise) {
       throw new NotFoundException('존재하지 않는 상품입니다.');
     }
 
-    //작성한 유저 id 일치한지 확인 로직 추가 필요
+    // product 생성자와 수정 요청한 사용자가 일치한지 확인
+    if (merchandise.manager.userId !== userId) {
+      throw new ForbiddenException('수정 권한이 없습니다.');
+    }
 
     if (title !== undefined) {
       merchandise.title = title;
@@ -254,20 +270,20 @@ export class MerchandiseService {
   }
 
   //상품 삭제 API
-  async remove(merchandiseId: number) {
+  async remove(merchandiseId: number, userId: number) {
     const merchandise = await this.merchandiseRepository.findOne({
       where: { id: merchandiseId },
-      relations: ['merchandiseImage', 'merchandiseOption'],
+      relations: ['merchandiseImage', 'merchandiseOption', 'manager'],
     });
 
     if (!merchandise) {
       throw new NotFoundException('존재하지 않는 상품입니다.');
     }
 
-    //작성 유저인지 확인
-    // if (merchandise.product.user.Id !== userId) {
-    //   throw new ForbiddenException('권한이 없습니다');
-    // }
+    // product 생성자와 수정 요청한 사용자가 일치한지 확인
+    if (merchandise.manager.userId !== userId) {
+      throw new ForbiddenException('수정 권한이 없습니다.');
+    }
 
     await this.merchandiseRepository.delete(merchandise.id);
